feat: add the download-mainnet-snapshot command to snapshot-compatibility
